requirements.txt
asgiref==3.6.0
backports.zoneinfo==0.2.1
certifi==2022.12.7
charset-normalizer==3.0.1
Django==4.1.5
django-parler==2.3
django-rosetta==0.9.8
idna==3.4
Pillow==9.4.0
polib==1.1.1
python-decouple==3.7
PyYAML==6.0
requests==2.28.2
sqlparse==0.4.3
typing-extensions==4.4.0
urllib3==1.26.14
